requirements.txt
asgiref==3.6.0
certifi==2022.12.7
charset-normalizer==3.0.1
crispy-bootstrap5==0.7
Django==4.1.5
django-appconf==1.0.5
django-ckeditor==6.5.1
django-crispy-forms==1.14.0
django-filter==22.1
django-js-asset==2.0.0
django-staticfiles==1.2.1
djangorestframework==3.14.0
idna==3.4
Pillow==9.4.0
postgres==4.0
psycopg2-binary==2.9.5
psycopg2-pool==1.1
pytz==2022.7.1
requests==2.28.2
sqlparse==0.4.3
tzdata==2022.7
urllib3==1.26.14
